+++ src/torrent.rs
@@ -35,7 +35,7 @@ async fn get_announce_response(
         .query(&[("port", 12457)])
         .query(&[("uploaded", 0)])
         .query(&[("downloaded", 0)])
-        .query(&[("left", meta.pieces.iter().map(|p| p.length).sum::<u32>())])
+        .query(&[("left", meta.pieces.iter().map(|p| p.length).sum::<usize>())])
         .build()
         .unwrap();
     let res = client.execute(req).await.unwrap().bytes().await.unwrap();
